Create Archipelago
Contains the game board.
getIslandName returns the name of the current island (or sea).
setSail makes the captain set sail.
getSpeed returns the speed of the captain's current ship.
getIslandNames returns a list of the island names.
travelTo makes the captain dock at the given island.
island_ returns the price of _ on the current island.
wallet returns the amount of doubloons the Captain has.
cargoSpace returns the space in use on the Captain's ship.
cargoTotal returns the total space on the Captain's ship.
getMaxItem returns the max amount of an item the Captain can purchase.
buyItem buys a number of items.
sellItem sells a number of items.
costItem returns the cost of selling a number of items.
captainItem returns the number of item onboard.
crewMemberAvailable returns true if the current island crew isn't part
of the crew.
getCrewName returns the name of the island's crew member.
getCrewCost returns the cost of the island's crew member.
hireCrew hires the crew member of the current island.
addCargoMember sets a boolean as true in all ships.
addNavigator sets a boolean as true in all ships.
addShipwright sets a boolean as true in all ships.
shipDescirptions returns all ships description.
shipDescription returns the current ship's description.
buyShip buys the chosen ship.